Applying Filter to Contacts, Users, Roles and Logs

+++ resources/views/roles/index.blade.php
@@ -12,6 +12,10 @@
 @include('layouts/error')
 @include('layouts/success')
 
+@include('layouts/search', ['url' => route('roles.index'),
+    'placeholder' => 'Enter with Role Name'
+])
+
 
 <table class="table table-bordered">
     <thead>
@@ -22,33 +26,41 @@
         </tr>
     </thead>
     <tbody>
-    @foreach ($roles as $key => $role)
-    <tr class="row m-0">
-        <td class="d-inline-block col-1">{{ $role->id }}</td>
-        <td class="d-inline-block col-9">{{ $role->name }}</td>
-        <td class="d-inline-block col-2 actions">
-            <a class="text-secondary" href="{{ route('roles.show',$role->id) }}"><i class="fas fa-search"></i></a>
-            @can('role-edit')
-                <a class="text-secondary" href="{{ route('roles.edit',$role->id) }}"><i class="fas fa-pen"></i></a>
-            @endcan
-            @if($role->id  != 1)
-                @can('role-delete')
-                    {!! Form::open(['method' => 'DELETE','route' => ['roles.destroy', $role->id],'id'=>'form-delete'.$role->id]) !!}
-                     <a class="text-danger" href="javascript:void(0)" onclick="$('#form-delete{{$role->id}}').submit();">
-                        <i class="fas fa-trash"></i>
-                      </a>
-                    {!! Form::close() !!}
-                @endcan
-            @else
-                @can('role-delete')
-                      <a class="text-danger disabled" href="javascript:void(0)">
-                        <i class="fas fa-ban"></i>
-                      </a>
-                @endcan
-            @endif
-        </td>
-    </tr>
-    @endforeach
+        @if ($roles->count() > 0)
+            @foreach ($roles as $key => $role)
+            <tr class="row m-0">
+                <td class="d-inline-block col-1">{{ $role->id }}</td>
+                <td class="d-inline-block col-9">{{ $role->name }}</td>
+                <td class="d-inline-block col-2 actions">
+                    <a class="text-secondary" href="{{ route('roles.show',$role->id) }}"><i class="fas fa-search"></i></a>
+                    @can('role-edit')
+                        <a class="text-secondary" href="{{ route('roles.edit',$role->id) }}"><i class="fas fa-pen"></i></a>
+                    @endcan
+                    @if($role->id  != 1)
+                        @can('role-delete')
+                            {!! Form::open(['method' => 'DELETE','route' => ['roles.destroy', $role->id],'id'=>'form-delete'.$role->id]) !!}
+                            <a class="text-danger" href="javascript:void(0)" onclick="$('#form-delete{{$role->id}}').submit();">
+                                <i class="fas fa-trash"></i>
+                            </a>
+                            {!! Form::close() !!}
+                        @endcan
+                    @else
+                        @can('role-delete')
+                            <a class="text-danger disabled" href="javascript:void(0)">
+                                <i class="fas fa-ban"></i>
+                            </a>
+                        @endcan
+                    @endif
+                </td>
+            </tr>
+            @endforeach
+        @else
+            <tr>
+                <td class="text-center d-inline-block col-12">
+                    No results found to: <strong>{{$search}}</strong>
+                </td>
+            </tr>
+        @endif
     </tbody>
 </table>
